% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DatetoWY}
\alias{DatetoWY}
\title{DatetoWY}
\usage{
DatetoWY(x)
}
\arguments{
\item{x}{\link{vector} vector of date strings}
}
\value{
Takes a Date Column and Returns The October-start Water Year for the Dates
}
\description{
DatetoWY
}
