% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixSiteDF}
\alias{fixSiteDF}
\title{fixSiteDF}
\usage{
fixSiteDF(d, site = "SITE_NO")
}
\arguments{
\item{d}{dataframe to be returned}

\item{site}{string indicating the name of the site column to fix}
}
\value{
Takes a dataframe and Fixes a Column of USGS 8-digit Site Numbers
}
\description{
fixSiteDF
}
