% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fixSiteDFread}
\alias{fixSiteDFread}
\title{fixSiteDFread}
\usage{
fixSiteDFread(file, site = "SITE_NO")
}
\arguments{
\item{file}{file path to read in the dataframe}

\item{site}{string indicating the name of the site column to fix}
}
\value{
Reads in a dataframe and Fixes a Column of USGS 8-digit Site Numbers
}
\description{
fixSiteDFread
}
