% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{posixtodec}
\alias{posixtodec}
\title{posixtodec}
\usage{
posixtodec(x)
}
\arguments{
\item{x}{\link{vector} vector of date strings}
}
\value{
Takes a Date Column and Returns Decimal-Year
}
\description{
posixtodec
}
