% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readQfile}
\alias{readQfile}
\title{readQfile}
\usage{
readQfile(site, Qdir)
}
\arguments{
\item{site}{8-digit USGS site number}

\item{Qdir}{file directory where Q data is stored}
}
\value{
Finds a stored USGS Flow file or Downloads the Data if Not Found
}
\description{
readQfile
}
