% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readQfile_year}
\alias{readQfile_year}
\title{readQfile_year}
\usage{
readQfile_year(site, Qdir, retrieve = TRUE)
}
\arguments{
\item{site}{8-digit USGS site number}

\item{Qdir}{file directory where Q data is stored}
}
\value{
Finds an annually summed stored USGS Flow file or Downloads the Data if Not Found
}
\description{
readQfile_year
}
