% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readwqualfile}
\alias{readwqualfile}
\title{readwqualfile}
\usage{
readwqualfile(site, param, dir)
}
\arguments{
\item{site}{8-digit USGS site number}

\item{param}{5-digit USGS parameter code}

\item{dir}{file directory where data is stored}
}
\value{
Finds a stored USGS Water Quality data file or Downloads the Data if Not Found
}
\description{
readwqualfile
}
