% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieveQdatamonth}
\alias{retrieveQdatamonth}
\title{retrieveQdatamonth}
\usage{
retrieveQdatamonth(site, Qdir, WY = TRUE)
}
\arguments{
\item{site}{8-digit USGS site number}

\item{Qdir}{file directory where Q data is stored, inherited from readQfile_year}
}
\value{
Calculates Monthly Summed USGS Flow Data and Downloads Data If Needed
}
\description{
retrieveQdatamonth
}
