% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{retrieveQdatayear}
\alias{retrieveQdatayear}
\title{retrieveQdatayear}
\usage{
retrieveQdatayear(site, Qdir, runclimate = TRUE, climatepath)
}
\arguments{
\item{site}{8-digit USGS site number}

\item{Qdir}{file directory where Q data is stored, inherited from readQfile_year}
}
\value{
Calculates Annually Summed USGS Flow Data and Downloads Data If Needed
}
\description{
retrieveQdatayear
}
